+++ plugins/modules/change_request_info.py
@@ -29,7 +29,7 @@ extends_documentation_fragment:
   - servicenow.itsm.sys_id.info
   - servicenow.itsm.number.info
   - servicenow.itsm.query
-  - servicenow.itsm.mapping
+  - servicenow.itsm.change_request_mapping
 seealso:
   - module: servicenow.itsm.change_request
 """
@@ -252,7 +252,7 @@ def sysparms_query(module, table_client, mapper):
 
 
 def run(module, table_client, attachment_client):
-    mapper = get_mapper(module, "change_request", PAYLOAD_FIELDS_MAPPING)
+    mapper = get_mapper(module, "change_request_mapping", PAYLOAD_FIELDS_MAPPING)
 
     if module.params["query"]:
         query = {"sysparm_query": sysparms_query(module, table_client, mapper)}
@@ -274,7 +274,9 @@ def main():
     module = AnsibleModule(
         supports_check_mode=True,
         argument_spec=dict(
-            arguments.get_spec("instance", "sys_id", "number", "query", "mapping"),
+            arguments.get_spec(
+                "instance", "sys_id", "number", "query", "change_request_mapping"
+            ),
         ),
         mutually_exclusive=[("sys_id", "query"), ("number", "query")],
     )
