+++ plugins/modules/incident.py
@@ -31,7 +31,7 @@ extends_documentation_fragment:
   - servicenow.itsm.sys_id
   - servicenow.itsm.number
   - servicenow.itsm.attachments
-  - servicenow.itsm.mapping
+  - servicenow.itsm.incident_mapping
 options:
   state:
     description:
@@ -173,7 +173,7 @@ DIRECT_PAYLOAD_FIELDS = (
 
 
 def ensure_absent(module, table_client, attachment_client):
-    mapper = get_mapper(module, "incident", PAYLOAD_FIELDS_MAPPING)
+    mapper = get_mapper(module, "incident_mapping", PAYLOAD_FIELDS_MAPPING)
     query = utils.filter_dict(module.params, "sys_id", "number")
     incident = table_client.get_record("incident", query)
 
@@ -216,7 +216,7 @@ def validate_params(params, incident=None):
 
 
 def ensure_present(module, table_client, attachment_client):
-    mapper = get_mapper(module, "incident", PAYLOAD_FIELDS_MAPPING)
+    mapper = get_mapper(module, "incident_mapping", PAYLOAD_FIELDS_MAPPING)
 
     query = utils.filter_dict(module.params, "sys_id", "number")
     payload = build_payload(module, table_client)
@@ -282,7 +282,9 @@ def run(module, table_client, attachment_client):
 
 def main():
     module_args = dict(
-        arguments.get_spec("instance", "sys_id", "number", "attachments", "mapping"),
+        arguments.get_spec(
+            "instance", "sys_id", "number", "attachments", "incident_mapping"
+        ),
         state=dict(
             type="str",
             choices=[
