+++ plugins/modules/incident_info.py
@@ -28,7 +28,7 @@ extends_documentation_fragment:
   - servicenow.itsm.sys_id.info
   - servicenow.itsm.number.info
   - servicenow.itsm.query
-  - servicenow.itsm.mapping
+  - servicenow.itsm.incident_mapping
 seealso:
   - module: servicenow.itsm.incident
 """
@@ -217,7 +217,7 @@ def sysparms_query(module, table_client, mapper):
 
 
 def run(module, table_client, attachment_client):
-    mapper = get_mapper(module, "incident", PAYLOAD_FIELDS_MAPPING)
+    mapper = get_mapper(module, "incident_mapping", PAYLOAD_FIELDS_MAPPING)
 
     if module.params["query"]:
         query = {"sysparm_query": sysparms_query(module, table_client, mapper)}
@@ -241,7 +241,9 @@ def main():
     module = AnsibleModule(
         supports_check_mode=True,
         argument_spec=dict(
-            arguments.get_spec("instance", "sys_id", "number", "query", "mapping"),
+            arguments.get_spec(
+                "instance", "sys_id", "number", "query", "incident_mapping"
+            ),
         ),
         mutually_exclusive=[("sys_id", "query"), ("number", "query")],
     )
